LayoutTests/imported/w3c/web-platform-tests/css/css-values/container-progress-serialize.tentative-expected.txt


FAIL 'calc(container-progress(width from 0px to 1px) / 1000)' as a specified value should serialize as 'calc(container-progress(width from 0px to 1px) / 1000)'. assert_not_equals: 'calc(container-progress(width from 0px to 1px) / 1000)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(container-progress(width from 0px to 1px) / 1000))' as a specified value should serialize as 'scale(calc(container-progress(width from 0px to 1px) / 1000))'. assert_not_equals: 'scale(calc(container-progress(width from 0px to 1px) / 1000))' should be valid in transform. got disallowed value ""
FAIL 'calc(container-progress(width from 0px to 1px) / 1000)' as a computed value should serialize as '0.228'. assert_equals: 'calc(container-progress(width from 0px to 1px) / 1000)' and '0.228' should serialize the same in computed values. expected "0.228" but got "1"
FAIL 'scale(calc(container-progress(width from 0px to 1px) / 1000))' as a computed value should serialize as 'matrix(0.228, 0, 0, 0.228, 0, 0)'. assert_equals: 'scale(calc(container-progress(width from 0px to 1px) / 1000))' and 'matrix(0.228, 0, 0, 0.228, 0, 0)' should serialize the same in computed values. expected "matrix(0.228, 0, 0, 0.228, 0, 0)" but got "none"
FAIL 'calc(0.1 * container-progress(height of my-container from 0px to 10em))' as a specified value should serialize as 'calc(0.1 * container-progress(height of my-container from 0px to 10em))'. assert_not_equals: 'calc(0.1 * container-progress(height of my-container from 0px to 10em))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(0.1 * container-progress(height of my-container from 0px to 10em)))' as a specified value should serialize as 'scale(calc(0.1 * container-progress(height of my-container from 0px to 10em)))'. assert_not_equals: 'scale(calc(0.1 * container-progress(height of my-container from 0px to 10em)))' should be valid in transform. got disallowed value ""
FAIL 'calc(0.1 * container-progress(height of my-container from 0px to 10em))' as a computed value should serialize as '0.228'. assert_equals: 'calc(0.1 * container-progress(height of my-container from 0px to 10em))' and '0.228' should serialize the same in computed values. expected "0.228" but got "1"
FAIL 'scale(calc(0.1 * container-progress(height of my-container from 0px to 10em)))' as a computed value should serialize as 'matrix(0.228, 0, 0, 0.228, 0, 0)'. assert_equals: 'scale(calc(0.1 * container-progress(height of my-container from 0px to 10em)))' and 'matrix(0.228, 0, 0, 0.228, 0, 0)' should serialize the same in computed values. expected "matrix(0.228, 0, 0, 0.228, 0, 0)" but got "none"
